Tile's image creation refactoring - NOT YET USABLE

+++ iOS/G3MiOSSDK/Commons/Mesh/AbstractGeometryMesh.cpp
@@ -134,9 +134,10 @@ void AbstractGeometryMesh::createGLState() {
                                                0,            //Stride 0
                                                true,         //Depth test
                                                false, 0,
-                                               false, (float)0.0, (float)0.0,
+                                               false, 0, 0,
                                                _lineWidth,
-                                               true, _pointSize), false);
+                                               true, _pointSize),
+                         false);
 
   if (_normals != NULL) {
     _glState->addGLFeature(new VertexNormalGLFeature(_normals, 3, 0, false, 0),
@@ -144,7 +145,8 @@ void AbstractGeometryMesh::createGLState() {
   }
 
   if (_translationMatrix != NULL) {
-    _glState->addGLFeature(new ModelTransformGLFeature(_translationMatrix->asMatrix44D()), false);
+    _glState->addGLFeature(new ModelTransformGLFeature(_translationMatrix->asMatrix44D()),
+                           false);
   }
 }
